move types to separate file for test passing
