+++ src/components/Toast/Toast.tsx
@@ -2,24 +2,7 @@ import * as React from 'react';
 import styles from './Toast.st.css';
 import { TPAComponentsConsumer } from '../TPAComponentsConfig';
 import { CloseIconButton } from './CloseIconButton/CloseIconButton';
-
-export enum TOAST_SKIN {
-  status = 'status',
-  success = 'success',
-  error = 'error',
-}
-
-export interface ToastProps {
-  /** Define the styles for toast */
-  skin: TOAST_SKIN;
-  shouldShowCloseButton?: boolean;
-  /** Callback function, will be called when click on close button */
-  onClose?(event: MouseEvent): void;
-}
-
-export interface DefaultProps {
-  shouldShowCloseButton: boolean;
-}
+import { DefaultProps, ToastProps } from './types';
 
 export class Toast extends React.Component<ToastProps> {
   static displayName = 'Toast';
