+++ src/components/Toast/index.tsx
@@ -1,1 +1,2 @@
 export * from './Toast';
+export * from './types';
